refactor(github): combine SingleRepoForm and BatchRepoForm into RepoAnalysisForm
- Merge SingleRepoForm and BatchRepoForm into a single component called RepoAnalysisForm
- Update GitHub page to use the new RepoAnalysisForm component
- Remove unused Alert component from GitHub page
- Update README with new features and usage instructions
